LIB-103 엔티티 수정에 따른 "장바구니 삭제" 수정 (#9)
* refactor: 엔티티수정에 따른 '장바구니삭제' 수정

* rename ForbbidenExp to NotYourResourceExp

* refactor: use create method in MockFactory.createOrders

* rename worng method name

* rename methods & change visibility

+++ src/main/java/com/liberty52/product/service/applicationservice/CartItemRemoveServiceImpl.java
@@ -1,7 +1,8 @@
 package com.liberty52.product.service.applicationservice;
 
-import com.liberty52.product.global.exception.external.ForbiddenException;
+import com.liberty52.product.global.exception.external.NotYourResource;
 import com.liberty52.product.global.exception.external.ResourceNotFoundException;
+import com.liberty52.product.global.exception.external.UnRemovableResourceException;
 import com.liberty52.product.service.entity.CustomProduct;
 import com.liberty52.product.service.repository.CartItemRepository;
 import com.liberty52.product.service.repository.CustomProductOptionRepository;
@@ -15,15 +16,24 @@ import org.springframework.transaction.annotation.Transactional;
 public class CartItemRemoveServiceImpl implements CartItemRemoveService {
     private final CartItemRepository cartItemRepository;
     private final CustomProductOptionRepository productCartOptionRepository;
-    private static final String RESOURCE_NAME = "CartItem";
+    private static final String RESOURCE_NAME_CART_ITEM = "CartItem(CustomProduct in Cart)";
     private static final String PARAM_NAME_ID = "ID";
+    private static final String RESOURCE_NAME_ORDER_ITEM = "OrderItem(CustomProduct in Order)";
 
 
+    /**
+     *
+     * @param authId owner of cartItem(customProduct)
+     * @param cartItemId
+     */
     @Override
     public void removeCartItem(String authId, String cartItemId) {
-        CustomProduct cartItem = cartItemRepository.findById(cartItemId).orElseThrow(() -> new ResourceNotFoundException(RESOURCE_NAME, PARAM_NAME_ID, cartItemId));
+        CustomProduct cartItem = cartItemRepository.findById(cartItemId).orElseThrow(() -> new ResourceNotFoundException(RESOURCE_NAME_CART_ITEM, PARAM_NAME_ID, cartItemId));
         if(!authId.equals(cartItem.getAuthId()))
-            throw new ForbiddenException(RESOURCE_NAME, authId);
+            throw new NotYourResource(RESOURCE_NAME_CART_ITEM, authId);
+        if (cartItem.isInOrder()) {
+            throw new UnRemovableResourceException(RESOURCE_NAME_ORDER_ITEM, cartItemId);
+        }
         productCartOptionRepository.deleteAll(cartItem.getOptions());
         cartItemRepository.delete(cartItem);
     }
